Make a rectangle version of everyone's picture.
* This version still have some problem to iron out, but I expected to do that before the end of the week
* Some picture are very hard to find a good way to crop, for example: Rex's.
* Beside Justin and Fiona, most people's bio fit. Further work should be done to style the bio, and ensure the font for the whole page.

Also, we should create some kind of avatar for everyone. That should be fun :D

More works are also expected on seperating everyone's profile into individual files. There are some JavaScript to do in this area.

We also need a build system to assemble these. I hope that after I finished with working, I would have more time to work on the website and get it online before Splash.